+++ src/pools/selection-strategies/selection-strategies-types.ts
@@ -65,7 +65,7 @@ export interface WorkerChoiceStrategyOptions {
  *
  * @internal
  */
-export interface RequiredStatistics {
+export interface TaskStatistics {
   /**
    * Require tasks runtime.
    */
@@ -101,9 +101,9 @@ export interface RequiredStatistics {
  */
 export interface IWorkerChoiceStrategy {
   /**
-   * Required tasks usage statistics.
+   * Required tasks statistics.
    */
-  readonly requiredStatistics: RequiredStatistics
+  readonly taskStatistics: TaskStatistics
   /**
    * Resets strategy internals.
    *