+++ src/worker/thread-worker.ts
@@ -1,11 +1,8 @@
 import { type MessagePort, isMainThread, parentPort } from 'node:worker_threads'
-import type {
-  MessageValue,
-  TaskFunctions,
-  WorkerFunction
-} from '../utility-types'
+import type { MessageValue } from '../utility-types'
 import { AbstractWorker } from './abstract-worker'
 import type { WorkerOptions } from './worker-options'
+import type { TaskFunctions, WorkerFunction } from './worker-functions'
 
 /**
  * A thread worker used by a poolifier `ThreadPool`.